+++ src/app/shared/confirm-dialog/confirm-dialog-data.ts
@@ -1,0 +1,6 @@
+export interface ConfirmDialogData {
+    title?: string;
+    text: string;
+    confirmButtonText: string;
+    cancelButtonText: string;
+}