+++ src/app/upload/file-upload-status.enum.ts
@@ -1,0 +1,6 @@
+export enum FileUploadStatus {
+    Pending,
+    Success,
+    Error,
+    Progress
+}