+++ src/app/video/shared/video-section.model.ts
@@ -1,0 +1,8 @@
+import { Video } from './video.model';
+
+export class VideoSection {
+    title: string;
+    isDismissable: boolean;
+    videos: Video[];
+    avatarUrl: string;
+}