+++ src/app/video/shared/video.model.ts
@@ -1,0 +1,8 @@
+export class Video {
+    id: string;
+    thumbnailUrl: string;
+    title: string;
+    owner: string;
+    views: number;
+    publishDate: Date;
+}