+++ src/app/video/video-routing.module.ts
@@ -1,21 +1,16 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { VideosComponent } from './videos/videos.component';
 import { VideoComponent } from './video.component';
-import { VideoUploadComponent } from './video-upload/video-upload.component';
+import { VideoSectionListComponent } from './video-section-list/video-section-list.component';
 
 const routes: Routes = [
-  {
-    path: '',
-    component: VideosComponent
-  },
   {
     path: 'watch/:id',
     component: VideoComponent
   },
   {
-    path: 'upload',
-    component: VideoUploadComponent
+    path: '',
+    component: VideoSectionListComponent
   }
 ];