+++ src/app/video/video-tile/video-tile.component.ts
@@ -1,0 +1,19 @@
+import { Component, OnInit, Input } from '@angular/core';
+import { Video } from '../shared/video.model';
+
+@Component({
+  selector: 'bs-video-tile',
+  templateUrl: './video-tile.component.html',
+  styleUrls: ['./video-tile.component.scss']
+})
+export class VideoTileComponent implements OnInit {
+
+  @Input() horizontal = false;
+  @Input() video: Video;
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}